lineageos: refactor updater scripts
Instead of relying on the device_deps.json from hudson being up-to-date,
we get this information ourselves from the lineage.dependencies files in
the individual repos. When including source dirs in the robotnix build,
we follow this dependency mapping to include any required directories.

Additionally, the device-metadata.json fetcher has been extracted into
its own script to easily allow a user to modify the default set of
devices to be fetched by update-device-dirs.py