Use tag instead of revision for fetchgit
Vanilla has repos with multiple tags and revisions that refer to the
same contents. Passing the revision hash to fetchgit / nix-prefetch-git
would include it in the checked-out /nix/store path. If we used the
revision in fetchgit, we'd be checking out the same contents multiple
times. So, if the revisionExpr is already a tag, just pass that to
nix-prefetch-git / fetchgit.