seedvault: add includedInFlavor option
Previously, we have stripped out upstream versions of packages if we
have a corresponding robotnix module which has the same functionality.
To better match upstream builds for reproducibility, we should not do
this _unless_ the upstream package is a prebuilt APK that we want to
build ourselves.

Since Seedvault (as of Android 11) can be built inside AOSP, and does
not have to be included as a prebuilt APK, we can use the flavor's
version instead, if it exists.

We introduce an internal option called apps.seedvault.includedInFlavor,
which is meant to be set by the flavor, and indicates that Seedvault
would be included automatically by the flavor if no additional changes
are made.  The robotnix module is only used if `enable=true` and
`includedInFlavor=false` under `apps.seedvault`.

If the user really wants to use the robotnix-provided version, they'll
need to set:
```nix
{
  apps.seedvault.enable = true;
  apps.seedvault.includedInFlavor = false;

  # Also remove the upstream version
  removedProductPackages = [ "Seedvault" ];
  source.dirs."external/seedvault".enable = false;
}
```