Refactor and separate kernel building for vanilla and GrapheneOS
The two flavors are different enough that trying to have a single module
for both was a pain.  Additionally, to avoid regressions, we now use the
./build.sh script from the GrapheneOS repositories. This will ensure
we're using exactly the right defconfig, utilities in PATH, and
resulting kernel source name.

Separating these two modules is a prerequisite for #116 as well.