Fix improper dependency of generateKeysScript on keyStorePath
Building `generateKeysScript` with a simple configuration of:

{device="crosshatch"; flavor="vanilla"; signing.enable=true; }

would result in Nix trying to import the key fingerprint for chromium.
This is shouldn't happen because the generateKeysScript should have no
data dependency on the actual APK that would be built.

The chain of dependencies was generateKeysScript -> keyTools ->
config.source.dirs -> nonNullPrebuilts (in modules/apps/prebuilts).

This nonNullPrebuilts would have to further evaluate the apk to decide
if it was null, which would ultimately pull depend on things further
down the chain (such as the IFD step to get the cert fingerprint).

Fixed by simply adding a apps.prebuilt.enable option, which defaults to
true if the apps.prebuilt.<name> submodule is set.

Closes #135