source: Don't dereference links when linking
This ensures overridden symlinks to directories are not attempted to be
linked as a subdir by `ln`.

e.g. with `build/core` already a symlink to something, the following
would not work:

```
   $ mkdir -p build
   $ ln -sf --relative build/make/core build/core
ln: failed to create symbolic link 'build/core/core': Permission denied
```

Observe how the destination is `build/core`, but the link attempted to
be created is `build/core/core`.

From the manpage:

> ```
>     --no-dereference treat LINK_NAME as a normal file if it is a
>                      symbolic link to a directory
> ```