release: include python dependency unconditionally
When generating the `ota` output, brillo_update_payload may invoke
python if a partition size is not divisible by 4096 bytes.  This is true
even for Android 12 (not only on older versions).

This only happens sometimes. For instance, truncate_file/python is not
invoked for `{ device="marlin"; flavor="lineageos"; }`, but it is for
`{ device="pioneer"; flavor="lineageos"; }`