Provide recoveryImg output
This is useful for devices that have a dedicated recovery image. This
includes, but is not limited to old devices, and *newest* devices. This
is because newest devices with dynamic partitions has dropped the silly
"boot as recovery" scheme, and rely again on a discrete recovery
partition.

This is useful mainly for LineageOS like deployment schemes.