lineageos: Lineage 21.0 support
Regarding the chooscombo to breakfast change on LOS: It's a wrapper around lunch
and I'm sure that somehow calls choosecombo or something. It takes a device as
parameter which is much nicer and selects the currently supported Android
release automatically which appears to be required for LOS21.

Originally this commit added support/workarounds for weird devices but this was
dropped because we don't support all LOS devices anymore.

MicroG doesn't need a patch anymore because lineageos allows the upstream µG
binary to spoof signatures by default. The binaries must be signed by the
upstream keys however, hence making them presigned.

Co-authored-by: Atemu <git@atemu.net>