check whether linkfile destination already exists in unpack script
This is relevant for the debug environment scripts use case - we want to
be able to run them on the same directory tree multiple times. However,
`ln` was attempting to create a symlink inside the Nix store if the
actual one already exists, as the path to the symlink was erroneously
interpreted as the directory of the symlink in that case. For instance,
the invocation

ln -sf --relative build/make/core build/core

attempted to create the symlink `build/core/core` if the symlink
`build/core -> build/make` already exists.