adevtool: use --noVerify flag
In commit d12ea01878da2db2b8841155ae736436bd91389b, GrapheneOS adevtool
introduced a setting which made the generated vendor dir Makefiles fail
if the git invocation is unsuccessful. I think it's sensible to just
flip that off instead of patching out the check or to pass a fake git
to `lunch` too.